requirements.txt
Flask==2.3.2
requests>=2.31
python-dotenv>=1.0.0
schedule==1.2.1
rich>=13.5 